Added .gitignore Removed IQCaptureButton and replaced with DBCameraButton - Removed IQ_SCSiriWaveformView and replaced with it's CocoaPods Replaced podspec with podspec.json

+++ IQMediaPickerController/IQAssetsPickerController/IQAlbumAssetsViewController.m
@@ -114,7 +114,7 @@
         
         if (self.assetController.maximumItemCount > 0)
         {
-            finalText = [finalText stringByAppendingFormat:@" (%lu maximum) ",self.assetController.maximumItemCount];
+            finalText = [finalText stringByAppendingFormat:@" (%lu maximum) ",(unsigned long)self.assetController.maximumItemCount];
         }
         self.selectedMediaCountItem.title = finalText;
     }
